requirements.txt
streamlit>=1.39.0
requests>=2.31.0
pandas>=2.0.0
yfinance>=0.2.18
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.8.0
google-generativeai>=0.3.0
pytest
scipy
matplotlib
ib_insync>=0.9.86
discord.py>=2.3.0
pyqlib